moving MetricsObserver out of internal
