+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/options/NavigationOptions.kt
@@ -16,7 +16,6 @@ data class NavigationOptions constructor(
     val fasterRouteDetectorInterval: Long,
     val distanceFormatter: DistanceFormatter?,
     val onboardRouterConfig: MapboxOnboardRouterConfig?,
-    val mapboxNavigationVersionName: String = "1.0-SNAPSHOT",
     val isFromNavigationUi: Boolean = false
 ) {
 
@@ -39,7 +38,6 @@ data class NavigationOptions constructor(
         private var fasterRouteDetectorInterval: Long = DEFAULT_FASTER_ROUTE_DETECTOR_INTERVAL,
         private var distanceFormatter: DistanceFormatter? = null,
         private var onboardRouterConfig: MapboxOnboardRouterConfig? = null,
-        private var mapboxNavigationVersionName: String = "1.0-SNAPSHOT",
         private var isFromNavigationUi: Boolean = false
     ) {
 
@@ -61,9 +59,6 @@ data class NavigationOptions constructor(
         fun onboardRouterConfig(onboardRouterConfig: MapboxOnboardRouterConfig?) =
             apply { this.onboardRouterConfig = onboardRouterConfig }
 
-        fun versionName(name: String) =
-                apply { this.mapboxNavigationVersionName = name }
-
         fun sdkName(flag: Boolean) =
                 apply { this.isFromNavigationUi = flag }
 
@@ -75,7 +70,6 @@ data class NavigationOptions constructor(
                 fasterRouteDetectorInterval,
                 distanceFormatter,
                 onboardRouterConfig,
-                mapboxNavigationVersionName,
                 isFromNavigationUi
             )
         }
