+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/MapboxNavigation.kt
@@ -61,6 +61,7 @@ private const val MAPBOX_NAVIGATION_USER_AGENT_BASE = "mapbox-navigation-android
 private const val MAPBOX_NAVIGATION_UI_USER_AGENT_BASE = "mapbox-navigation-ui-android"
 private const val MAPBOX_NAVIGATION_TOKEN_EXCEPTION_OFFBOARDROUTER = "You need to provide an token access in order to use the default OffboardRouter."
 private const val MAPBOX_NAVIGATION_TOKEN_EXCEPTION_ONBOARDROUTER = "You need to provide an token access in order to use the default OnboardRouter."
+private const val MAPBOX_NAVIGATION_TOKEN_EXCEPTION = "A valid token is required"
 
 /**
  * ## Mapbox Navigation Core SDK
@@ -178,7 +179,7 @@ constructor(
             }
         ifNonNull(accessToken) { token ->
             MapboxMetricsReporter.init(context, accessToken
-                ?: throw RuntimeException(MAPBOX_NAVIGATION_TOKEN_EXCEPTION_OFFBOARDROUTER), obtainUserAgent(navigationOptions.isFromNavigationUi))
+                ?: throw RuntimeException(MAPBOX_NAVIGATION_TOKEN_EXCEPTION), obtainUserAgent(navigationOptions.isFromNavigationUi))
             MapboxNavigationTelemetry.initialize(
                 context.applicationContext,
                 token,
@@ -539,7 +540,7 @@ constructor(
         return if (isFromNavigationUi) {
             "$MAPBOX_NAVIGATION_UI_USER_AGENT_BASE/${BuildConfig.MAPBOX_NAVIGATION_VERSION_NAME}"
         } else {
-            "{MAPBOX_NAVIGATION_USER_AGENT_BASE}/${BuildConfig.MAPBOX_NAVIGATION_VERSION_NAME}"
+            "$MAPBOX_NAVIGATION_USER_AGENT_BASE/${BuildConfig.MAPBOX_NAVIGATION_VERSION_NAME}"
         }
     }
